feat: add asset configuration and npm installation logic
